+++ src/WR.cpp
@@ -1,7 +1,7 @@
 #include "WR.h"
 using namespace std;
 
-WR::WR(const string &code, int id, const string &reservoir, const string &municipality, int maxDelivery)
+WR::WR(const string &code, int id, const string &reservoir, const string &municipality, double maxDelivery)
         : WaterElement(code, id) {
     this->reservoir = reservoir;
     this->municipality = municipality;
@@ -24,10 +24,10 @@ void WR::setMunicipality(const string &municipality) {
     WR::municipality = municipality;
 }
 
-int WR::getMaxDelivery() const {
+double WR::getMaxDelivery() const {
     return maxDelivery;
 }
 
-void WR::setMaxDelivery(int maxDelivery) {
+void WR::setMaxDelivery(double maxDelivery) {
     WR::maxDelivery = maxDelivery;
 }
